+++ torchutils/trainer/__init__.py
@@ -1,2 +1,2 @@
 from .engine import Trainer
-from torchutils.utils.pydantic import TrainerModel
+from torchutils.utils.pydantic import TrainerModel, TrainerModelBuilder